Some refactoring and fixing bug in TaskTracker
Moving code to process batches into Aggregator2 (which will be renamed). Part
of the reason for this refactoring is logging. Info or debug level logging can
be turned on for AggregationManager without flooding the log file with a lot
of log statements.

TaskTracker provides synchronization for schedules tasks. It is used to track
when all of the scheduled tasks have completed. CountDownLatch cannot be used
anymore because it requires you to specify the number of tasks or events up
front. This worked fine with the metrics_index table since we were only
querying against a single row, and we would iterate over the result set to get
the count. It is not know in advance though how many rows there will be in
metrics_cache so a different sychronization mechanism was needed.

Lastly I should mention that I am not referring to CQL rows here but rather
rows at the actual storage layer, i.e., thrift rows.