cleaning up and re-enabling some tests
This commit also updates the logic for determine when aggregation last ran
after a server outage. It is a bit more complicated now since with
metrics_cache there can be multiple partitions per time slice; so, we have to
iterate over all possible partitiion keys per time slice, and if we fail to
find any data, only then do we check the previous time slice. It was eaasier
with the metrics_index table because there was only one partition per time
slice.