changing naming convention for the bucket column in metrics_cache
Previously when inserting a row into metrics_index, the bucket column would
have the next wider bucket as its value. For example, if we inserted raw data,
then we would update metrics_index with the bucket column having a value of
"one_hour_metrics". With the schema changes to start using metrics_cache in
place of metrics_index, this approach got confusing. The bucket column now
is set to the name of the historical table in which data is being written.