update cache and cache index only when necessary for late data
When aggregating late data whose 6 hour time slice has not yet finished, we
should update metrics_cache and metrics_cache_index so that the necessary rows
are present when 6 hour data aggregation runs. If the 6 hour time slice has
already passed, then there is no need to update metrics_cache and
metrics_cache_index because we go ahead and recompute the 6 hour data. The
same idea holds with 24 hour data. If the 24 hour time slice has not yet
finished, then should update metrics_cache and metrics_cache_index so that the
necessary rows are present when 24 hour data aggregation runs; however, if the
24 hour time slice has already passed, then there is no need to perform the
updates because we will (in the next commit) go ahead and recompute 24 hour
data.